requirements.txt
affine
astroid
attrs
autopep8
backports.csv
certifi
chardet
click
click-plugins
cligj
colorama
defusedxml
diff-match-patch
django
django-crispy-forms
django-import-export
django-leaflet
et_xmlfile
gdal
geos
hdf4
hdf5
icc_rt
icu
idna
intel-openmp
isort
jdcal
jpeg
kealib
krb5
lazy-object-proxy
libboost
libcurl
libgdal
libiconv
libkml
libnetcdf
libpng
libpq
libspatialite
libssh2
libtiff
libxml2
mccabe
mkl
mkl_fft
mkl_random
numpy
numpy-base
oauthlib
odfpy
openpyxl
openssl
pcre
pip
proj4
psycopg2
pycodestyle
pyjwt
pylint
pyparsing
python
python3-openid
pytz
pyyaml
rasterio
requests
requests-oauthlib
setuptools
shapely
six
snuggs
social-auth-app-django
social-auth-core
sqlite
sqlparse
tablib
tk
typed-ast
urllib3
vc
vs2015_runtime
wheel
wincertstore
wrapt
xerces-c
xlrd
xlwt
xz
yaml
zlib
zstd
